+++ Vantage/Animation2D/Commands/ICommand.cs
@@ -1,18 +1,12 @@
-// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="ICommand.cs" company="">
-//   
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
-
 namespace Vantage.Animation2D.Commands
 {
     public interface ICommand
     {
         #region Public Properties
 
-        float EndTime { get; set; }
+        double EndTime { get; set; }
 
-        float StartTime { get; set; }
+        double StartTime { get; set; }
 
         #endregion
