+++ Vantage/Animation2D/Commands/VScaleCommand.cs
@@ -1,12 +1,10 @@
 ﻿namespace Vantage.Animation2D.Commands
 {
-    using System.Globalization;
-
     using Vantage.Animation2D.OsbTypes;
 
     public class VScaleCommand : Command<OsbScale>
     {
-        public VScaleCommand(int easing, float startTime, float endTime, OsbScale startValue, OsbScale endValue)
+        public VScaleCommand(int easing, double startTime, double endTime, OsbScale startValue, OsbScale endValue)
             : base("V", easing, startTime, endTime, startValue, endValue)
         {
         }
