+++ Vantage/Animation3D/Animation/EasingCurves/IEasingCurve.cs
@@ -2,6 +2,6 @@ namespace Vantage.Animation3D.Animation.EasingCurves
 {
     public interface IEasingCurve
     {
-        float Evaluate(float x);
+        double Evaluate(double x);
     }
 }
