+++ Vantage/Animation3D/Animation/Keyframes/DiscreteKeyframe.cs
@@ -2,17 +2,17 @@
 {
     public class DiscreteKeyframe<T> : IKeyframe<T>
     {
-        public DiscreteKeyframe(float time, T value)
+        public DiscreteKeyframe(double time, T value)
         {
             this.Time = time;
             this.Value = value;
         }
 
-        public float Time { get; set; }
+        public double Time { get; set; }
 
         public T Value { get; set; }
 
-        public T ValueAtTime(IKeyframe<T> next, float time)
+        public T ValueAtTime(IKeyframe<T> next, double time)
         {
             return this.Value;
         }
